+++ build-logic/convention/src/main/kotlin/KotlinLibraryConventionPlugin.kt
@@ -31,6 +31,7 @@ class KotlinLibraryConventionPlugin : Plugin<Project> {
             }
 
             configureDetekt()
+            configureTestLogging()
 
             dependencies {
                 add("testImplementation", libs.findBundle("unit-test").get())